requirements.txt
ollama
langchain
langchain-community
chromadb
sentence-transformers
pdfplumber
nltk
rouge-score
rich
pandas
pyyaml
tiktoken
numpy

